fix: use same code in notebooks and file inputs for image upload
