+++ frontend/src/lib/lemon-ui/LemonFileInput/LemonFileInput.tsx
@@ -4,7 +4,6 @@ import { LemonTag } from 'lib/lemon-ui/LemonTag/LemonTag'
 import { Spinner } from 'lib/lemon-ui/Spinner/Spinner'
 import clsx from 'clsx'
 import './LemonFileInput.scss'
-import api from 'lib/api'
 
 export interface LemonFileInputProps extends Pick<HTMLInputElement, 'multiple' | 'accept'> {
     value?: File[]
@@ -18,62 +17,6 @@ export interface LemonFileInputProps extends Pick<HTMLInputElement, 'multiple' |
      * styling is applied to the alternativeDropTargetRef
      * **/
     alternativeDropTargetRef?: RefObject<HTMLElement>
-}
-
-/**
- * The library used relies on canvas.toBlob() which has slightly odd behaviour
- * It tends to convert things to png unexpectedly :'(
- * See http://kangax.github.io/jstests/toDataUrl_mime_type_test/ for a test that shows this behavior
- */
-function canReduceThisBlobType(file: File): boolean {
-    const supportedTypes = ['image/png', 'image/jpeg', 'image/webp']
-    return supportedTypes.includes(file.type)
-}
-
-export function useUploadFiles({
-    onUpload,
-    onError,
-}: {
-    onUpload?: (url: string, fileName: string) => void
-    onError: (detail: string) => void
-}): {
-    setFilesToUpload: (files: File[]) => void
-    filesToUpload: File[]
-    uploading: boolean
-} {
-    const [uploading, setUploading] = useState(false)
-    const [filesToUpload, setFilesToUpload] = useState<File[]>([])
-    useEffect(() => {
-        const uploadFiles = async (): Promise<void> => {
-            if (filesToUpload.length === 0) {
-                setUploading(false)
-                return
-            }
-
-            try {
-                setUploading(true)
-                const formData = new FormData()
-                let file: File = filesToUpload[0]
-                if (canReduceThisBlobType(file)) {
-                    const compressedBlob = await lazyImageBlobReducer(file)
-                    file = new File([compressedBlob], file.name, { type: compressedBlob.type })
-                }
-
-                formData.append('image', file)
-                const media = await api.media.upload(formData)
-                onUpload?.(media.image_location, media.name)
-            } catch (error) {
-                const errorDetail = (error as any).detail || 'unknown error'
-                onError(errorDetail)
-            } finally {
-                setUploading(false)
-                setFilesToUpload([])
-            }
-        }
-        uploadFiles().catch(console.error)
-    }, [filesToUpload])
-
-    return { setFilesToUpload, filesToUpload, uploading }
 }
 
 export const LemonFileInput = ({
@@ -203,11 +146,6 @@ export const LemonFileInput = ({
             </div>
         </>
     )
-}
-
-export const lazyImageBlobReducer = async (blob: Blob): Promise<Blob> => {
-    const blobReducer = (await import('image-blob-reduce')).default()
-    return blobReducer.toBlob(blob, { max: 2000 })
 }
 
 function acceptToDisplayName(accept: string): string {
